avoid cpp see thrust
