+++ common/unified/base/kernel_launch.hpp
@@ -17,6 +17,7 @@
 #if defined(GKO_COMPILING_CUDA)
 
 #define GKO_KERNEL __device__
+#include "common/cuda_hip/base/math.hpp"
 #include "common/cuda_hip/base/types.hpp"
 
 
@@ -43,6 +44,7 @@ GKO_INLINE GKO_ATTRIBUTES constexpr unpack_member_type<T> unpack_member(T value)
 #elif defined(GKO_COMPILING_HIP)
 
 #define GKO_KERNEL __device__
+#include "common/cuda_hip/base/math.hpp"
 #include "common/cuda_hip/base/types.hpp"
