+++ cuda/base/config.hpp
@@ -6,9 +6,8 @@
 #define GKO_CUDA_BASE_CONFIG_HPP_
 
 
+#include <ginkgo/core/base/math.hpp>
 #include <ginkgo/core/base/types.hpp>
-
-#include "common/cuda_hip/base/math.hpp"
 
 
 namespace gko {
