+++ cuda/base/cublas_bindings.hpp
@@ -10,7 +10,6 @@
 
 #include <ginkgo/core/base/exception_helpers.hpp>
 
-#include "common/cuda_hip/base/math.hpp"
 #include "common/cuda_hip/base/types.hpp"
 
 
@@ -227,23 +226,6 @@ template <typename ValueType>
 GKO_BIND_CUBLAS_NORM2(ValueType, detail::not_implemented);
 
 #undef GKO_BIND_CUBLAS_NORM2
-
-
-inline cublasHandle_t init(cudaStream_t stream)
-{
-    cublasHandle_t handle;
-    GKO_ASSERT_NO_CUBLAS_ERRORS(cublasCreate(&handle));
-    GKO_ASSERT_NO_CUBLAS_ERRORS(
-        cublasSetPointerMode(handle, CUBLAS_POINTER_MODE_DEVICE));
-    GKO_ASSERT_NO_CUBLAS_ERRORS(cublasSetStream(handle, stream));
-    return handle;
-}
-
-
-inline void destroy(cublasHandle_t handle)
-{
-    GKO_ASSERT_NO_CUBLAS_ERRORS(cublasDestroy(handle));
-}
 
 
 }  // namespace cublas
