+++ hip/base/hipblas_bindings.hip.hpp
@@ -240,24 +240,6 @@ GKO_BIND_HIPBLAS_NORM2(ValueType, detail::not_implemented);
 #undef GKO_BIND_HIPBLAS_NORM2
 
 
-inline hipblasContext* init(hipStream_t stream)
-{
-    hipblasHandle_t handle;
-    GKO_ASSERT_NO_HIPBLAS_ERRORS(hipblasCreate(&handle));
-    GKO_ASSERT_NO_HIPBLAS_ERRORS(
-        hipblasSetPointerMode(handle, HIPBLAS_POINTER_MODE_DEVICE));
-    GKO_ASSERT_NO_HIPBLAS_ERRORS(hipblasSetStream(handle, stream));
-    return reinterpret_cast<hipblasContext*>(handle);
-}
-
-
-inline void destroy_hipblas_handle(hipblasContext* handle)
-{
-    GKO_ASSERT_NO_HIPBLAS_ERRORS(
-        hipblasDestroy(reinterpret_cast<hipblasHandle_t>(handle)));
-}
-
-
 }  // namespace hipblas
